refactor: extract HumanoidRobot base class; remove duplication across H1/H1_2/G1 env files
The three humanoid robot environments (H1, H1_2, G1) were near-complete
copies of each other — 10 methods, ~120 lines, were pasted verbatim in
every file. The only code that legitimately differed was the hip-joint
DOF indices used in _reward_hip_pos().

Changes
-------
legged_gym/envs/base/humanoid_robot.py  (new file)
  Introduces HumanoidRobot(LeggedRobot) — a shared base class that owns:
  - _init_foot() / _init_buffers()      (rigid-body foot state setup)
  - update_feet_state()                 (per-step foot state refresh)
  - _post_physics_step_callback()       (gait phase computation)
  - _get_noise_scale_vec()              (observation noise scaling)
  - compute_observations()              (obs + privileged_obs assembly)
  - _reward_contact()                   (phase-gated contact reward)
  - _reward_feet_swing_height()         (swing foot height penalty)
  - _reward_alive()                     (constant survival reward)
  - _reward_contact_no_vel()            (contact-while-moving penalty)
  - _reward_hip_pos()                   (uses self.hip_indices — set per subclass)

legged_gym/envs/h1/h1_env.py
legged_gym/envs/h1_2/h1_2_env.py
legged_gym/envs/g1/g1_env.py
  Each reduced to a 3-line subclass that sets hip_indices.
  Zero behaviour change — the logic is identical, just deduplicated.

legged_gym/scripts/play.py
  Removed duplicate imports of sys and LEGGED_GYM_ROOT_DIR (lines 1-5
  contained each import twice). Removed unused RECORD_FRAMES and
  MOVE_CAMERA variables (set in __main__ but never read anywhere).

legged_gym/envs/base/legged_robot_config.py + legged_gym/utils/terrain.py
  Corrected long-standing spelling mistake: slope_treshold → slope_threshold
  (both the config field definition and the call site in terrain.py).
