+++ legged_gym/envs/base/legged_robot_config.py
@@ -34,7 +34,7 @@ class LeggedRobotCfg(BaseConfig):
         # terrain types: [smooth slope, rough slope, stairs up, stairs down, discrete]
         terrain_proportions = [0.1, 0.1, 0.35, 0.25, 0.2]
         # trimesh only:
-        slope_treshold = 0.75 # slopes above this threshold will be corrected to vertical surfaces
+        slope_threshold = 0.75 # slopes above this threshold will be corrected to vertical surfaces
 
     class commands:
         curriculum = False
